// src/app/store/action-types/product-action-types.ts
export enum ProductsActionTypes {
  GET_PRODUCTS_DATA = '[Products] Get products data',
  GET_PRODUCTS_SUCCESS = '[Products] Get products success',
  GET_PRODUCTS_FAILURE = '[Products] Get products failure',

  DELETE_PRODUCT = '[Cart] Delete product',
  DELETE_PRODUCT_SUCCESS = '[Cart] Delete product success',
  DELETE_PRODUCT_FAILURE = '[Cart] Delete product failure',

  UPDATE_PRODUCT = '[Cart] Update product',
  UPDATE_PRODUCT_SUCCESS = '[Cart] Update product success',
  UPDATE_PRODUCT_FAILURE = '[Cart] Update product failure',
}